Add API for polling controller input
The controller buttons on the GB have some unpredictable settle times so
we need to be careful to only read from the controller some time after
selecting the input source.

Rather than trying to schedule this, lets just read the input state once
per frame and save it into a convenient location.