Add a high-level state machine manager
This will allow strongly-typed state machines to be implemented.
I'm hoping that this will be useful to structure game mode state,
animations, and sound effects.

Warning: this is broken in the current toolchain release... I need to
come up with a high performance calling convention that does not break
indirect function calls.